requirements.txt
xtralien
pyserial
numpy
pandas
matplotlib
PyQt6
pyqtgraph
streamlit
plotly
